TimespanLogging Functions

TimespanLogging.get_logs!

TimespanLogging.init_similar

TimespanLogging.timespan_finish

TimespanLogging.timespan_start

Basic Functions

TimespanLogging.timespan_start — Function

timespan_start(ctx, category::Symbol, id, tl)

Generates an Event{:start} which denotes the start of an event. The event is categorized by category, and uniquely identified by id; these two must be the same passed to timespan_finish to close the event. tl is the "timeline" of the event, which is just an arbitrary payload attached to the event.

TimespanLogging.timespan_finish — Function

timespan_finish(ctx, category::Symbol, id, tl)

Generates an Event{:finish} which denotes the end of an event. The event is categorized by category, and uniquely identified by id; these two must be the same as previously passed to timespan_start. tl is the "timeline" of the event, which is just an arbitrary payload attached to the event.

TimespanLogging.get_logs! — Function

get_logs!(::LocalEventLog, raw=false; only_local=false) -> Union{Vector{Timespan},Vector{Event}}

Get the logs from each process' local event log, clearing it in the process. Set raw to true to get potentially unmatched Events; the default is to return only matched events as Timespans. If only_local is set true, only process-local logs will be fetched; the default is to fetch logs from all processes.

Logging Metric Functions

TimespanLogging.init_similar — Function

Creates a copy of x with the same configuration, but fresh/empty data.