TimespanLogging Types

TimespanLogging.Event

TimespanLogging.Events.CPULoadAverages

TimespanLogging.Events.CoreMetrics

TimespanLogging.Events.DebugMetrics

TimespanLogging.Events.EventSaturation

TimespanLogging.Events.FullMetrics

TimespanLogging.Events.IDMetrics

TimespanLogging.Events.LogWindow

TimespanLogging.Events.MemoryFree

TimespanLogging.Events.TimelineMetrics

TimespanLogging.LocalEventLog

TimespanLogging.MultiEventLog

TimespanLogging.NoOpLog

Log Sink Types

TimespanLogging.MultiEventLog — Type

MultiEventLog

Processes events immediately, generating multiple log streams. Multiple consumers may register themselves in the MultiEventLog, and when accessed, log events will be provided to all consumers. A consumer is simply a function or callable struct which will be called with an event when it's generated. The return value of the consumer will be pushed into a log stream dedicated to that consumer. Errors thrown by consumers will be caught and rendered, but will not otherwise interrupt consumption by other consumers, or future consumption cycles. An error will result in nothing being appended to that consumer's log.

TimespanLogging.LocalEventLog — Type

LocalEventLog

Stores events in a process-local array. Accessing the logs is all-or-nothing; if multiple consumers call get_logs!, they will get different sets of logs.

TimespanLogging.NoOpLog — Type

NoOpLog

Disables event logging entirely.

Event Types

TimespanLogging.Event — Type

An event generated by timespan_start or timespan_finish.

Built-in Event Types

TimespanLogging.Events.CoreMetrics — Type

CoreMetrics

Tracks the timestamp, category, and kind of the Event object generated by log events.

TimespanLogging.Events.IDMetrics — Type

IDMetrics

Tracks the ID of Event objects generated by log events.

TimespanLogging.Events.TimelineMetrics — Type

TimelineMetrics

Tracks the timeline of Event objects generated by log events.

TimespanLogging.Events.FullMetrics — Type

FullMetrics

Tracks the full Event object generated by log events.

TimespanLogging.Events.CPULoadAverages — Type

CPULoadAverages

Monitors the CPU load averages.

TimespanLogging.Events.MemoryFree — Type

MemoryFree

Monitors the percentage of free system memory.

TimespanLogging.Events.EventSaturation — Type

EventSaturation

Tracks the compute saturation (running tasks) per-processor.

TimespanLogging.Events.DebugMetrics — Type

Debugging metric, used to log event start/finish via @debug.

TimespanLogging.Events.LogWindow — Type

LogWindow

Aggregator that prunes events to within a given time window.

```