Dagger Types

Dagger.AnyScope

Dagger.ArrayDomain

Dagger.Blocks

Dagger.Chunk

Dagger.Context

Dagger.DArray

Dagger.DTask

Dagger.Deps

Dagger.Events.BytesAllocd

Dagger.Events.ProcessorSaturation

Dagger.Events.WorkerSaturation

Dagger.ExactScope

Dagger.In

Dagger.InOut

Dagger.NodeScope

Dagger.OSProc

Dagger.Options

Dagger.Out

Dagger.ProcessScope

Dagger.Processor

Dagger.Sch.SchedulerOptions

Dagger.Sch.ThunkOptions

Dagger.Shard

Dagger.TaintScope

Dagger.ThreadProc

Dagger.Thunk

Dagger.UnionScope

Dagger.UnitDomain

Dagger.ProcessorTypeScope

Task Types

Dagger.Thunk — Type

Thunk

Wraps a callable object to be run with Dagger. A Thunk is typically created through a call to delayed or its macro equivalent @par.

Constructors

delayed(f; kwargs...)(args...)
@par [option=value]... f(args...)

Examples

julia> t = delayed(sin)(π)  # creates a Thunk to be computed later
Thunk(sin, (π,))

julia> collect(t)  # computes the result and returns it to the current process
1.2246467991473532e-16

Arguments

f: The function to be called upon execution of the Thunk.

args: The arguments to be passed to the Thunk.

kwargs: The properties describing unique behavior of this Thunk. Details

for each property are described in the next section.

option=value: The same as passing kwargs to delayed.

Public Properties

meta::Bool=false: If true, instead of fetching cached arguments from

Chunks and passing the raw arguments to f, instead pass the Chunk. Useful for doing manual fetching or manipulation of Chunk references. Non-Chunk arguments are still passed as-is.

processor::Processor=OSProc() - The processor associated with f. Useful if

f is a callable struct that exists on a given processor and should be transferred appropriately.

scope::Dagger.AbstractScope=DefaultScope() - The scope associated with f.

Useful if f is a function or callable struct that may only be transferred to, and executed within, the specified scope.

Options

options: A Sch.ThunkOptions struct providing the options for the Thunk.

If omitted, options can also be specified by passing key-value pairs as kwargs.

Dagger.DTask — Type

DTask

Returned from Dagger.@spawn/Dagger.spawn calls. Represents a task that is in the scheduler, potentially ready to execute, executing, or finished executing. May be fetch'd or wait'd on at any time. See Dagger.@spawn for more details.

Task Options Types

Dagger.Options — Type

Options(::NamedTuple)
Options(; kwargs...)

Options for thunks and the scheduler. See Task Spawning for more information.

Dagger.Sch.ThunkOptions — Type

ThunkOptions

Stores Thunk-local options to be passed to the Dagger.Sch scheduler.

Arguments

single::Int=0: (Deprecated) Force thunk onto worker with specified id. 0 disables this option.

proclist=nothing: (Deprecated) Force thunk to use one or more processors that are instances/subtypes of a contained type. Alternatively, a function can be supplied, and the function will be called with a processor as the sole argument and should return a Bool result to indicate whether or not to use the given processor. nothing enables all default processors.

time_util::Dict{Type,Any}: Indicates the maximum expected time utilization for this thunk. Each keypair maps a processor type to the utilization, where the value can be a real (approximately the number of nanoseconds taken), or MaxUtilization() (utilizes all processors of this type). By default, the scheduler assumes that this thunk only uses one processor.

alloc_util::Dict{Type,UInt64}: Indicates the maximum expected memory utilization for this thunk. Each keypair maps a processor type to the utilization, where the value is an integer representing approximately the maximum number of bytes allocated at any one time.

occupancy::Dict{Type,Real}: Indicates the maximum expected processor occupancy for this thunk. Each keypair maps a processor type to the utilization, where the value can be a real between 0 and 1 (the occupancy ratio, where 1 is full occupancy). By default, the scheduler assumes that this thunk has full occupancy.

allow_errors::Bool=true: Allow this thunk to error without affecting non-dependent thunks.

checkpoint=nothing: If not nothing, uses the provided function to save the result of the thunk to persistent storage, for later retrieval by restore.

restore=nothing: If not nothing, uses the provided function to return the (cached) result of this thunk, were it to execute. If this returns a Chunk, this thunk will be skipped, and its result will be set to the Chunk. If nothing is returned, restoring is skipped, and the thunk will execute as usual. If this function throws an error, restoring will be skipped, and the error will be displayed.

storage::Union{Chunk,Nothing}=nothing: If not nothing, references a MemPool.StorageDevice which will be passed to MemPool.poolset internally when constructing Chunks (such as when constructing the return value). The device must support MemPool.CPURAMResource. When nothing, uses MemPool.GLOBAL_DEVICE[].

storage_root_tag::Any=nothing: If not nothing, specifies the MemPool storage leaf tag to associate with the thunk's result. This tag can be used by MemPool's storage devices to manipulate their behavior, such as the file name used to store data on disk."

storage_leaf_tag::MemPool.Tag,Nothing}=nothing: If not nothing, specifies the MemPool storage leaf tag to associate with the thunk's result. This tag can be used by MemPool's storage devices to manipulate their behavior, such as the file name used to store data on disk."

storage_retain::Bool=false: The value of retain to pass to MemPool.poolset when constructing the result Chunk.

Dagger.Sch.SchedulerOptions — Type

SchedulerOptions

Stores DAG-global options to be passed to the Dagger.Sch scheduler.

Arguments

single::Int=0: (Deprecated) Force all work onto worker with specified id. 0 disables this option.

proclist=nothing: (Deprecated) Force scheduler to use one or more processors that are instances/subtypes of a contained type. Alternatively, a function can be supplied, and the function will be called with a processor as the sole argument and should return a Bool result to indicate whether or not to use the given processor. nothing enables all default processors.

allow_errors::Bool=true: Allow thunks to error without affecting non-dependent thunks.

checkpoint=nothing: If not nothing, uses the provided function to save the final result of the current scheduler invocation to persistent storage, for later retrieval by restore.

restore=nothing: If not nothing, uses the provided function to return the (cached) final result of the current scheduler invocation, were it to execute. If this returns a Chunk, all thunks will be skipped, and the Chunk will be returned. If nothing is returned, restoring is skipped, and the scheduler will execute as usual. If this function throws an error, restoring will be skipped, and the error will be displayed.

Data Management Types

Dagger.Chunk — Type

Chunk

A reference to a piece of data located on a remote worker. Chunks are typically created with Dagger.tochunk(data), and the data can then be accessed from any worker with collect(::Chunk). Chunks are serialization-safe, and use distributed refcounting (provided by MemPool.DRef) to ensure that the data referenced by a Chunk won't be GC'd, as long as a reference exists on some worker.

Each Chunk is associated with a given Dagger.Processor, which is (in a sense) the processor that "owns" or contains the data. Calling collect(::Chunk) will perform data movement and conversions defined by that processor to safely serialize the data to the calling worker.

Constructors

See tochunk.

Dagger.Shard — Type

Maps a value to one of multiple distributed "mirror" values automatically when used as a thunk argument. Construct using @shard or shard.

Data Dependencies Types

Dagger.In — Type

Specifies a read-only dependency.

Dagger.Out — Type

Specifies a write-only dependency.

Dagger.InOut — Type

Specifies a read-write dependency.

Dagger.Deps — Type

Specifies one or more dependencies.

Processor Types

Dagger.Processor — Type

Processor

An abstract type representing a processing device and associated memory, where data can be stored and operated on. Subtypes should be immutable, and instances should compare equal if they represent the same logical processing device/memory. Subtype instances should be serializable between different nodes. Subtype instances may contain a "parent" Processor to make it easy to transfer data to/from other types of Processor at runtime.

Dagger.OSProc — Type

OSProc <: Processor

Julia CPU (OS) process, identified by Distributed pid. The logical parent of all processors on a given node, but otherwise does not participate in computations.

Dagger.ThreadProc — Type

ThreadProc <: Processor

Julia CPU (OS) thread, identified by Julia thread ID.

Scope Types

Dagger.AnyScope — Type

Widest scope that contains all processors.

Dagger.NodeScope — Type

Scoped to the same physical node.

Dagger.ProcessScope — Type

Scoped to the same OS process.

Dagger.ProcessorTypeScope — Function

Scoped to any processor with a given supertype.

Dagger.TaintScope — Type

Taints a scope for later evaluation.

Dagger.UnionScope — Type

Union of two or more scopes.

Dagger.ExactScope — Type

Scoped to a specific processor.

Context Types

Dagger.Context — Type

Context(xs::Vector{OSProc}) -> Context
Context(xs::Vector{Int}) -> Context

Create a Context, by default adding each available worker.

It is also possible to create a Context from a vector of OSProc, or equivalently the underlying process ids can also be passed directly as a Vector{Int}.

Special fields include:

'log_sink': A log sink object to use, if any.

profile::Bool: Whether or not to perform profiling with Profile stdlib.

Array Types

Dagger.DArray — Type

DArray{T,N,F}(domain, subdomains, chunks, concat)
DArray(T, domain, subdomains, chunks, [concat=cat])

An N-dimensional distributed array of element type T, with a concatenation function of type F.

Arguments

T: element type

domain::ArrayDomain{N}: the whole ArrayDomain of the array

subdomains::AbstractArray{ArrayDomain{N}, N}: a DomainBlocks of the same dimensions as the array

chunks::AbstractArray{Union{Chunk,Thunk}, N}: an array of chunks of dimension N

concat::F: a function of type F. concat(x, y; dims=d) takes two chunks x and y and concatenates them along dimension d. cat is used by default.

Dagger.Blocks — Type

Blocks(xs...)

Indicates the size of an array operation, specified as xs, whose length indicates the number of dimensions in the resulting array.

Dagger.ArrayDomain — Type

ArrayDomain{N}

An N-dimensional domain over an array.

Dagger.UnitDomain — Type

UnitDomain

Default domain – has no information about the value

Logging Event Types

Dagger.Events.BytesAllocd — Type

BytesAllocd

Tracks memory allocated for Chunks.

Dagger.Events.ProcessorSaturation — Type

ProcessorSaturation

Tracks the compute saturation (running tasks) per-processor.

Dagger.Events.WorkerSaturation — Type

WorkerSaturation

Tracks the compute saturation (running tasks).